+++ src/features/persons/family_members/index.tsx
@@ -1,51 +1,17 @@
 import { Box, Divider, Typography } from '@mui/material';
 
-import { useAtomValue } from 'jotai/react';
-import { personCurrentDetailsState, personsActiveState } from '@states/persons';
 import MemberSelector from './member_selector';
-import { useCallback, useMemo, useState } from 'react';
-import { setPersonCurrentDetails } from '@services/states/persons';
+import useFamilyMembers from './useFamilyMembers';
 
 const FamilyMembers = () => {
-	const personsActive = useAtomValue(personsActiveState);
-	const currentPerson = useAtomValue(personCurrentDetailsState);
-	const [newlyAddedMemberCount, setNewlyAddedMemberCount] = useState(1);
-	const familyMembers = useMemo(() => {
-		if (currentPerson.person_data.family_members) {
-			return currentPerson.person_data.family_members.members;
-		}
-		return [];
-	}, [currentPerson]);
-
-	const selectOptions = useMemo(
-		() =>
-			personsActive.filter(
-				(person) =>
-					!familyMembers.includes(person.person_uid) &&
-					person.person_uid !== currentPerson.person_uid
-			),
-		[currentPerson.person_uid, familyMembers, personsActive]
-	);
-
-	const handleAddNewMember = useCallback(() => {
-		setNewlyAddedMemberCount((prev) => prev + 1);
-	}, []);
-
-	const onSelectPerson = useCallback(
-		(personId: string) => {
-			const person = structuredClone(currentPerson);
-			if (person.person_data.family_members.members.includes(personId)) {
-				return;
-			}
-			person.person_data.family_members.members.push(personId);
-			person.person_data.family_members.updatedAt = new Date().toISOString();
-			setPersonCurrentDetails(person);
-			setNewlyAddedMemberCount((prev) => Math.max(prev - 1, 0));
-		},
-		[currentPerson]
-	);
-
-	console.log(familyMembers);
+	const {
+		handleAddNewMember,
+		newlyAddedMemberCount,
+		onSelectPerson,
+		options,
+		personsActive,
+		familyMembers,
+	} = useFamilyMembers();
 
 	return (
 		<Box
@@ -71,7 +37,9 @@ const FamilyMembers = () => {
 					label="Family member"
 					selected={member}
 					options={personsActive.filter(
-						(person) => person.person_uid === member || !familyMembers.includes(person.person_uid)
+						(person) =>
+							person.person_uid === member ||
+							!familyMembers.includes(person.person_uid)
 					)}
 					onSelectPerson={onSelectPerson}
 				/>
@@ -80,7 +48,7 @@ const FamilyMembers = () => {
 				<MemberSelector
 					key={idx}
 					label="Family member"
-					options={selectOptions}
+					options={options}
 					isLast={newlyAddedMemberCount === idx + 1}
 					onAddMember={handleAddNewMember}
 					onSelectPerson={onSelectPerson}
